Refactor CSS mirroring to script so the Windows build can access it
