+++ electron/build_action.sh
@@ -40,6 +40,9 @@ rsync -ac \
 # In Electron, the path is relative to electron_index.html.
 scripts/environment_json.sh -p windows > $OUTPUT/www/environment.json
 
+# Generate CSS rules to mirror the UI in RTL languages.
+node -e "require('./scripts/generate_rtl_css.js')('www/ui_components/*.html', '$OUTPUT/www/ui_components')"
+
 # We need a top-level index.js.
 # Its only job is to load electron/index.js.
 cat << EOM > $OUTPUT/index.js
